requirements.txt
asgiref==3.3.4
astunparse==1.6.3
attrs==21.2.0
autopep8==1.5.7
coverage==5.5
Django==3.2.2
djangorestframework==3.12.4
gunicorn==20.1.0
iniconfig==1.1.1
Jinja2==2.11.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
packaging==20.9
pdoc3==0.9.2
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.11.1
pytest-django==4.2.0
pytest-pythonpath==0.7.3
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
